initial metadata implementation
