+++ apps/console/src/app/(auth)/invite/page.tsx
@@ -1,8 +1,11 @@
-'use client'
-
 import React, { Suspense } from 'react'
 import { LoaderCircle } from 'lucide-react'
 import { InviteAccepter } from '@/components/pages/invite/accept'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Invite',
+}
 
 const AcceptInvite: React.FC = () => {
   return (
