+++ apps/console/src/app/(auth)/layout.tsx
@@ -1,4 +1,12 @@
+import { Metadata } from 'next'
 import { AuthLayout, type AuthLayoutProps } from '../../components/layouts/auth'
+
+export const metadata: Metadata = {
+  title: {
+    template: '%s | Openlane | Streamlining Compliance Securing Success',
+    default: 'Acme',
+  },
+}
 
 export default function Layout({ children }: AuthLayoutProps): React.ReactNode {
   return <AuthLayout>{children}</AuthLayout>
