+++ apps/console/src/app/(auth)/resend-verify/page.tsx
@@ -1,7 +1,10 @@
-'use client'
-
 import { Resend } from '@/components/pages/auth/resend-verify/resend'
 import { pageStyles } from './page.styles'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Resend Verify',
+}
 
 const ResendVerification: React.FC = () => {
   const { content, form } = pageStyles()
