+++ apps/console/src/app/(auth)/subscriber-verify/page.tsx
@@ -1,8 +1,11 @@
-'use client'
-
 import { LoaderCircle } from 'lucide-react'
 import React, { Suspense } from 'react'
 import { TokenVerifier } from '@/components/pages/auth/subscriber-verify/verifier'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Subscriber verify',
+}
 
 const VerifySubscriber: React.FC = () => {
   return (
