+++ apps/console/src/app/(protected)/controls/page.tsx
@@ -1,18 +1,13 @@
-'use client'
-
-import React, { useContext, useEffect } from 'react'
+import React from 'react'
 import { PageHeading } from '@repo/ui/page-heading'
 import ControlsTable from '@/components/pages/protected/controls/table/controls-table'
-import { BreadcrumbContext } from '@/providers/BreadcrumbContext'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Controls',
+}
 
 const Page: React.FC = () => {
-  const { setCrumbs } = useContext(BreadcrumbContext)
-  useEffect(() => {
-    setCrumbs([
-      { label: 'Home', href: '/dashboard' },
-      { label: 'Controls', href: '/controls' },
-    ])
-  }, [setCrumbs])
   return (
     <>
       <PageHeading heading="Controls" />
