+++ apps/console/src/app/(protected)/evidence/page.tsx
@@ -1,20 +1,13 @@
-'use client'
-
-import React, { useContext, useEffect } from 'react'
+import React from 'react'
 import EvidencePage from '@/components/pages/protected/evidence/evidence-page'
 import { PageHeading } from '@repo/ui/page-heading'
-import { BreadcrumbContext } from '@/providers/BreadcrumbContext'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Evidence',
+}
 
 const Page: React.FC = () => {
-  const { setCrumbs } = useContext(BreadcrumbContext)
-
-  useEffect(() => {
-    setCrumbs([
-      { label: 'Home', href: '/dashboard' },
-      { label: 'Evidence', href: '/evidence' },
-    ])
-  }, [setCrumbs])
-
   return (
     <>
       <PageHeading heading="Submit Evidence" />
