+++ apps/console/src/app/(protected)/groups/page.tsx
@@ -1,7 +1,13 @@
-'use client'
 import GroupsPage from '@/components/pages/protected/groups/groups-page'
 import React from 'react'
+import { Metadata } from 'next'
 
-const Page: React.FC = () => <GroupsPage />
+export const metadata: Metadata = {
+  title: 'Groups',
+}
+
+const Page: React.FC = () => {
+  return <GroupsPage />
+}
 
 export default Page
