+++ apps/console/src/app/(protected)/layout.tsx
@@ -2,8 +2,10 @@ import type { Metadata } from 'next'
 import { DashboardLayout } from '@/components/layouts/dashboard/dashboard'
 
 export const metadata: Metadata = {
-  title: 'Dashboard',
-  description: 'The open source foundation of a sustainable digital world',
+  title: {
+    template: 'Acme Corp: | %s',
+    default: 'Acme',
+  },
 }
 
 export default function RootLayout({ children }: { children: React.ReactNode }): React.ReactNode {
