+++ apps/console/src/app/(protected)/organization-settings/general-settings/page.tsx
@@ -5,7 +5,7 @@ import { pageStyles } from './page.styles'
 import { OrganizationDelete } from '@/components/pages/protected/organization/general-settings/organization-delete'
 
 export const metadata: Metadata = {
-  title: 'Organization settings',
+  title: 'General settings',
 }
 
 const Page: React.FC = () => {
