+++ apps/console/src/app/(protected)/questionnaires/page.tsx
@@ -1,24 +1,15 @@
-'use client'
-
-import React, { useContext, useEffect } from 'react'
 import { PageHeading } from '@repo/ui/page-heading'
 import { QuestionnairesTable } from '@/components/pages/protected/questionnaire/table/questionnaire-table.tsx'
-import { BreadcrumbContext } from '@/providers/BreadcrumbContext'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Questionnaires',
+}
 
 const Page: React.FC = () => {
-  const { setCrumbs } = useContext(BreadcrumbContext)
-
-  useEffect(() => {
-    setCrumbs([
-      { label: 'Home', href: '/dashboard' },
-      { label: 'Questionnaires', href: '/questionnaires' },
-    ])
-  }, [setCrumbs])
-
   return (
     <>
       <PageHeading heading="Questionnaires" />
-
       <QuestionnairesTable />
     </>
   )
