+++ apps/console/src/app/(protected)/user-settings/developers/page.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from 'next/types'
 import DevelopersPage from '@/components/pages/protected/developers/developers-user-page'
 
 export const metadata: Metadata = {
-  title: 'Organization settings',
+  title: 'Developers',
 }
 
 const Page: React.FC = () => {
