+++ apps/console/src/components/pages/protected/risks/view/view-risks-page.tsx
@@ -73,6 +73,7 @@ const ViewRisksPage: React.FC<TRisksPageProps> = ({ riskId }) => {
         stakeholderID: risk.stakeholder?.id,
         delegateID: risk.delegate?.id,
       })
+      document.title = `Acme Corp: Risks - ${risk.name}`
     }
   }, [risk])
